refactor: 탭 아이템 상수들 constants/tab.ts로 이동
