+++ src/constants/event.ts
@@ -1,8 +1,5 @@
 import { EventTags } from '@/types/event';
 
-import { PATH } from './path';
-import { TAB_CONSTANTS } from './tab';
-
 const APPLIED_EVENTS_TAGS: EventTags = {
   pending: {
     title: '관리자 확인중',
@@ -42,12 +39,6 @@ const APPLIED_EVENTS_TAGS: EventTags = {
   },
 };
 
-const MAIN_TABS = [
-  { title: `${TAB_CONSTANTS.PERFORMANCE}`, link: `${PATH.MAIN}` },
-  { title: `${TAB_CONSTANTS.EVENT_SHOW}`, link: `${PATH.MAIN_EVENT}` },
-  { title: `${TAB_CONSTANTS.RECRUITMENT}`, link: `${PATH.MAIN_RECRUITMENT}` },
-];
-
 const EVENT_DETAIL = {
   date: '날짜',
   time: '시간',
@@ -63,4 +54,4 @@ const EVENT_DETAIL_BUTTON = {
   apply: '참여 신청하기',
 };
 
-export { APPLIED_EVENTS_TAGS, EVENT_DETAIL, EVENT_DETAIL_BUTTON, MAIN_TABS };
+export { APPLIED_EVENTS_TAGS, EVENT_DETAIL, EVENT_DETAIL_BUTTON };
