+++ src/constants/tab.ts
@@ -1,3 +1,5 @@
+import { PATH } from './path';
+
 enum TAB_CONSTANTS {
   PERFORMANCE = '공연',
   EVENT_SHOW = '행사',
@@ -11,4 +13,15 @@ enum TAB_CONSTANTS {
   CLUB_MANAGE = '클럽 관리',
 }
 
-export { TAB_CONSTANTS };
+const MAIN_TABS = [
+  { title: `${TAB_CONSTANTS.PERFORMANCE}`, link: `${PATH.MAIN}` },
+  { title: `${TAB_CONSTANTS.EVENT_SHOW}`, link: `${PATH.MAIN_EVENT}` },
+  { title: `${TAB_CONSTANTS.RECRUITMENT}`, link: `${PATH.MAIN_RECRUITMENT}` },
+];
+
+const PROFILE_TABS = [
+  { title: `${TAB_CONSTANTS.APPLIED_EVENT}`, link: `${PATH.PROFILE}` },
+  { title: `${TAB_CONSTANTS.BOOKMARKED_EVENT}`, link: `${PATH.BOOKMARK}` },
+];
+
+export { TAB_CONSTANTS, MAIN_TABS, PROFILE_TABS };
